Pin doctor egress recovery regression
- Exercise provider-specific recovery through the public doctor output path.\n\nRef #1813
